Ticket #2869 Refactor experiment page Experiment chart data moved out from the ExperimentResultAdapter - DoubleIndexIterator moved to atlas-utils
